Add |newBoardOptions| function to |ChessBoard| class.
This takes a look at a potienal new sport for the knight and returns
a int8_t telling you how many options the knight will have when he gets
there. It takes into account the rules the knight plays by along with
checking to make sure the knight had not been there previously.